Show a few exemplar workflows for processing
Note that a couple common use cases for this sort of image processing
work. Namely there are use cases where people are batching through a lot
of image data, which is just looping over it more or less. Alternatively
users may want to process a single massive image, which they must break
up and work on pieces of and then glue the results back together.